implement LogoutCLI & update test cases
